feat : 얼굴인식 페이지 reloading, navigation -> mount 버그 해결
feat : 얼굴인식 페이지 reloading, navigation -> mount 버그 해결
