+++ OkeyDokeyContest/App.js
@@ -18,7 +18,7 @@ import Welcome from './src/screens/Welcome';
 import Identify from './src/screens/Identify';
 import Favorites from './src/screens/Favorites';
 import WebviewContainer from './src/pages/WebviewContainer';
-import FaceRecognition from './src/pages/FaceRecognition';
+import FaceRecognition from './src/pages/FaceReognition';
 const Stack = createNativeStackNavigator();
 
 const App = () => {
@@ -27,7 +27,7 @@ const App = () => {
       <Provider store={store}>
         <Stack.Navigator screenOptions={{headerShown: false}}>
           {/* <Stack.Screen name="MyWebViewScreen" component={MyWebViewScreen} /> */}
-          <Stack.Screen name="Welcome" component={Welcome} />
+          {/* <Stack.Screen name="Welcome" component={Welcome} />
           <Stack.Screen name="FaceRecognition" component={FaceRecognition} />
           <Stack.Screen name="Identify" component={Identify} />
           <Stack.Screen name="WebviewContainer" component={WebviewContainer} />
@@ -37,7 +37,7 @@ const App = () => {
           <Stack.Screen name="Qmilk" component={Qmilk} />
           <Stack.Screen name="EasyMenu" component={EasyMenu} />
           <Stack.Screen name="OrderCheck" component={OrderCheck} />
-          <Stack.Screen name="ShoppingBasket" component={ShoppingBasket} />
+          <Stack.Screen name="ShoppingBasket" component={ShoppingBasket} /> */}
           <Stack.Screen name="OrderNum" component={OrderNum} />
           <Stack.Screen name="Payment" component={Payment} />
           <Stack.Screen name="InputPhoneNum" component={InputPhoneNum} />
